+++ forui/test/src/widgets/popover/popover_test.dart
@@ -180,6 +180,9 @@ void main() {
       await tester.sendKeyEvent(LogicalKeyboardKey.tab);
       await tester.pumpAndSettle();
 
+      await tester.sendKeyEvent(LogicalKeyboardKey.tab);
+      await tester.pumpAndSettle();
+
       expect(Focus.of(tester.element(find.text('3'))).hasFocus, true);
     });
   });